refactor: move code from governance/classes to governance/superblock where it belongs
